Refactor CustomCSRs for extensibility beyond Rocket
