+++ src/main/scala/tile/CustomCSRs.scala
@@ -1,0 +1,42 @@
+// See LICENSE.SiFive for license details.
+
+package freechips.rocketchip.tile
+
+import chisel3._
+
+import freechips.rocketchip.config.Parameters
+
+case class CustomCSR(id: Int, mask: BigInt, init: Option[BigInt])
+
+object CustomCSR {
+  def constant(id: Int, value: BigInt): CustomCSR = CustomCSR(id, BigInt(0), Some(value))
+}
+
+class CustomCSRIO(implicit p: Parameters) extends CoreBundle {
+  val wen = Bool()
+  val wdata = UInt(xLen.W)
+  val value = UInt(xLen.W)
+}
+
+class CustomCSRs(implicit p: Parameters) extends CoreBundle {
+  // Not all cores have these CSRs, but those that do should follow the same
+  // numbering conventions.  So we list them here but default them to None.
+  protected def bpmCSRId = 0x7c0
+  protected def bpmCSR: Option[CustomCSR] = None
+
+  // If you override this, you'll want to concatenate super.decls
+  def decls: Seq[CustomCSR] = bpmCSR.toSeq
+
+  val csrs = Vec(decls.size, new CustomCSRIO)
+
+  def flushBTB = getOrElse(bpmCSR, _.wen, false.B)
+  def bpmStatic = getOrElse(bpmCSR, _.value(0), false.B)
+
+  protected def getByIdOrElse[T](id: Int, f: CustomCSRIO => T, alt: T): T = {
+    val idx = decls.indexWhere(_.id == id)
+    if (idx < 0) alt else f(csrs(idx))
+  }
+
+  protected def getOrElse[T](csr: Option[CustomCSR], f: CustomCSRIO => T, alt: T): T =
+    csr.map(c => getByIdOrElse(c.id, f, alt)).getOrElse(alt)
+}
